dom/nodes/Node-appendChild-script-and-iframe.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and iframe from a div</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
const happened = [];
const iframe = document.createElement("iframe");

test(() => {
  iframe.src = "data:text/html,"

  const script = document.createElement("script");
  script.textContent = `
    happened.push(iframe.contentWindow === null);
`;

  const div = document.createElement("div");
  div.appendChild(script);
  div.appendChild(iframe);

  assert_array_equals(happened, []);
  document.body.appendChild(div);
  assert_array_equals(happened, [false])
});
</script>
